move sidebar config in seperate file
