+++ documentation/src/libs/trim-trailing-slash.ts
@@ -1,0 +1,3 @@
+export function trimTrailingSlash(link: string) {
+  return link.endsWith("/") ? link.slice(0, -1) : link;
+}
